AG-138: delete dataset web app

+++ core/delete.py
@@ -2,13 +2,11 @@ import os
 import shutil
 
 from core.EsConnection import EsConnection
-from core.common import dataset_exists
+from core.common import dataset_exists, refresh_index
 
 
 def delete_file_by_name(dataset: str, file_name: str):
-    exists = dataset_exists(dataset)
-
-    if not exists:
+    if not dataset_exists(dataset):
         raise Exception(f"Dataset '{dataset}' does not exist.")
 
     es = EsConnection()
@@ -30,12 +28,11 @@ def delete_file_by_name(dataset: str, file_name: str):
     }
 
     es.connection.delete_by_query(index=es.index_name, query=query)
+    refresh_index()
 
 
 def delete_dataset_by_name(dataset: str):
-    exists = dataset_exists(dataset)
-
-    if not exists:
+    if not dataset_exists(dataset):
         raise Exception(f"Dataset '{dataset}' does not exist.")
 
     es = EsConnection()
@@ -46,13 +43,16 @@ def delete_dataset_by_name(dataset: str):
     }
 
     es.connection.delete_by_query(index=es.index_name, query=query)
+    refresh_index()
 
 
 def delete_dataset_files_by_name(dataset: str):
     folder_path = os.path.abspath(os.path.join(os.curdir, dataset))
-    shutil.rmtree(folder_path)
+    if os.path.exists(folder_path):
+        shutil.rmtree(folder_path)
 
 
 def delete_dataset_file_by_name(dataset: str, file_name: str):
     file_path = os.path.abspath(os.path.join(os.curdir, dataset, file_name))
-    os.remove(file_path)
+    if os.path.exists(file_path):
+        os.remove(file_path)
